# emmaus_ui/src/app/services/models/create-inquiry-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateInquiryRequest {
  message?: string;
  remark?: string;
  senderEmail?: string;
  senderName?: string;
  senderPhone?: string;
  status?: string;
  subject?: string;
  type?: string;
}