# emmaus_ui/src/app/services/models/document-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DocumentResponse {
  description?: string;
  displayOrder?: number;
  documentType?: string;
  publicId?: string;
  publishedDate?: string;
  thumbnailUrl?: string;
  title?: string;
  uploadedAt?: string;
  url?: string;
}