# emmaus_ui/src/app/services/models/page-response-document-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DocumentResponse } from '../models/document-response';
export interface PageResponseDocumentResponse {
  content?: Array<DocumentResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}