# emmaus_ui/src/app/services/models/page-response-inquiry-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InquiryResponse } from '../models/inquiry-response';
export interface PageResponseInquiryResponse {
  content?: Array<InquiryResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}