# emmaus_ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/inquiries/update/{public-id}": {
            "put": {
                "tags": [
                    "inquiries"
                ],
                "operationId": "updateInquiry",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInquiryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/s3/generate-presigned-url": {
            "post": {
                "tags": [
                    "s3"
                ],
                "operationId": "generatePresignedUrl",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneratePresignedUrlRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreSignedUrlDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/inquiries/create": {
            "post": {
                "tags": [
                    "inquiries"
                ],
                "operationId": "createInquiry",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInquiryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/documents/update/{public-id}": {
            "post": {
                "tags": [
                    "documents"
                ],
                "operationId": "updateDocument",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDocumentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/delete/{public-id}": {
            "post": {
                "tags": [
                    "documents"
                ],
                "operationId": "deleteDocument",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/documents/create": {
            "post": {
                "tags": [
                    "documents"
                ],
                "operationId": "createDocument",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDocumentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/s3/presigned-delete-url": {
            "get": {
                "tags": [
                    "s3"
                ],
                "operationId": "getPresignedUrlToDelete",
                "parameters": [
                    {
                        "name": "file_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "folder",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreSignedUrlDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/inquiries/pages": {
            "get": {
                "tags": [
                    "inquiries"
                ],
                "operationId": "getPagesOfInquiries",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseInquiryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/documents/pages": {
            "get": {
                "tags": [
                    "documents"
                ],
                "operationId": "getPagesOfDocuments",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "doc_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseDocumentResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateInquiryRequest": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
...
            "GeneratePresignedUrlRequest": {
                "type": "object",
                "properties": {
                    "fileName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "folder": {
                        "type": "string",
                        "minLength": 1
                    },
                    "contentType": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "contentType",
                    "fileName",
                    "folder"
                ]
            },
            "PreSignedUrlDto": {
                "type": "object",
                "properties": {
                    "preSignedUrl": {
                        "type": "string"
                    },
                    "publicUrl": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    }
                }
            },
            "CreateInquiryRequest": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "senderName": {
                        "type": "string"
                    },
                    "senderEmail": {
                        "type": "string"
                    },
                    "senderPhone": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
...
            "CreateDocumentRequest": {
                "type": "object",
                "properties": {
                    "documentType": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "thumbnailUrl": {
                        "type": "string"
                    },
                    "publishedDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "InquiryResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string"
                    },
                    "senderName": {
                        "type": "string"
                    },
                    "senderEmail": {
                        "type": "string"
                    },
                    "senderPhone": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PageResponseInquiryResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InquiryResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
            "DocumentResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "documentType": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "thumbnailUrl": {
                        "type": "string"
                    },
                    "publishedDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "uploadedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PageResponseDocumentResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}